requirements.txt
aiogram==2.25.1
psycopg2_binary==2.9.10
python-dotenv==1.0.1
